requirements.txt
appdirs==1.4.4
attrs==23.2.0
blinker==1.8.2
booleanOperations==0.9.0
Brotli==1.1.0
cffi==1.16.0
cffsubr==0.3.0
click==8.1.7
defcon==0.10.3
Flask==3.0.3
Flask-Cors==4.0.1
fontFeatures==1.8.0
fontMath==0.9.3
fonttools==4.51.0
freetype-py==2.4.0
fs==2.4.16
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.1
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pyclipper==1.3.0.post5
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.1
git+https://github.com/no-design-foundry/filters-rasterizer.git@03b642f629423c072d65055634fe63ac8bceba95#egg=rasterizer
git+https://github.com/no-design-foundry/filters-rotorizer.git@44a38feb37ed9b29bf40c8f3320d9c39aa57977b#egg=rotorizer
scipy==1.13.0
setuptools==69.5.1
six==1.16.0
tzdata==2024.1
ufo-extractor==0.6.0
ufo2ft==3.2.1
ufoLib2==0.16.0
uharfbuzz==0.39.1
unicodedata2==15.1.0
Werkzeug==3.0.3
xattr==1.1.0
zopfli==0.2.3
